GH-8 Refactor application folder structure
