+++ .gitignore
@@ -13,6 +13,7 @@
 
 # ide
 .idea/
+.vscode/
 
 # misc
 .DS_Store
