+++ src/components/App.tsx
@@ -5,8 +5,7 @@ import { Provider as ReduxProvider } from 'react-redux'
 import { ChakraProvider } from '@chakra-ui/react';
 
 import store from '../features/store';
-import AppHeader from './AppHeader';
-import AppErrorBox from './AppErrorBox';
+import AppHeader from './AppHeader/AppHeader';
 import AppNavigation from './AppNavigation';
 
 function App(): React.ReactElement {
@@ -15,7 +14,6 @@ function App(): React.ReactElement {
       <ChakraProvider>
         <HashRouter>
           <AppHeader />
-          <AppErrorBox />
           <AppNavigation />
         </HashRouter>
       </ChakraProvider>
