+++ src/components/AppError/AppErrorDetail.tsx
