+++ src/components/AppHeader/AppHeader.tsx
@@ -12,6 +12,7 @@ import {
 import { HamburgerIcon, CloseIcon } from '@chakra-ui/icons';
 
 import headerLogo from "../images/logo.png";
+
 import AppHeaderDesktop from './AppHeaderDesktop';
 import AppHeaderMobile from './AppHeaderMobile';
 import AppHeaderProfile from './AppHeaderProfile';
