+++ src/components/SurvivorPositionUpdate/SurvivorPositionUpdateFormFallback.tsx
@@ -1,9 +1,19 @@
 import React from 'react';
 
-import { Box, Flex, Skeleton, Text, useBreakpointValue } from '@chakra-ui/react';
+import {
+  Box,
+  Flex,
+  Skeleton,
+  Text,
+  useBreakpointValue
+} from '@chakra-ui/react';
 
 function SurvivorPositionUpdateFormFallback(): React.ReactElement {
-  const survivorPositionBoxWidth = useBreakpointValue({ base: '100%', md: '70%', lg: '60%' });
+  const survivorPositionBoxWidth = useBreakpointValue({
+    base: '100%',
+    md: '70%',
+    lg: '60%'
+  });
 
   return (
     <Flex direction={{ base: 'column' }} alignItems={'center'} gridGap={5}>
