+++ src/components/SurvivorRegistration/SurvivorRegistrationForm.tsx
@@ -18,9 +18,9 @@ import {
 } from '@chakra-ui/react';
 import { ViewIcon, ViewOffIcon } from '@chakra-ui/icons';
 
-import { GENDER_LIST } from '../constants/genderList';
-import { buildLocationString } from '../models/location';
-import { SurvivorGender } from '../models/survivor';
+import { GENDER_LIST } from '../../constants/genderList';
+import { buildLocationString } from '../../models/location';
+import { SurvivorGender } from '../../models/survivor';
 
 function SurvivorRegistrationForm(props: FormikProps<FormikValues>): React.ReactElement {
   const [showPassword, setShowPassword] = React.useState(false)
