+++ src/components/SurvivorRegistration/SurvivorRegistrationInventory.tsx
