+++ src/components/SurvivorRegistration/SurvivorRegistrationInventoryFallback.tsx
