+++ src/components/SurvivorRegistration/SurvivorRegistrationInventoryList.tsx
@@ -1,8 +1,15 @@
 import React from 'react';
-import { getIn, ArrayHelpers, FieldArray, FormikProps, FormikValues } from 'formik';
+import {
+  getIn,
+  ArrayHelpers,
+  FieldArray,
+  FormikProps,
+  FormikValues
+} from 'formik';
 
-import { ItemRead } from '../models/item';
-import { useGetItemsQuery } from '../services';
+import { ItemRead } from '../../models/item';
+import { useGetItemsQuery } from '../../services';
+
 import SurvivorRegistrationInventoryFallback from './SurvivorRegistrationInventoryFallback';
 import SurvivorRegistrationInventoryListItem from './SurvivorRegistrationInventoryListItem';
