+++ src/components/SurvivorRegistration/SurvivorRegistrationMapMarker.tsx
