+++ src/components/SurvivorRegistration/SurvivorRegistrationPage.tsx
@@ -3,12 +3,13 @@ import { Formik, FormikHelpers, FormikProps, FormikValues } from 'formik';
 
 import { Container, useToast } from '@chakra-ui/react';
 
-import { LEAFLET_DEFAULTS } from '../constants/leafletDefaults';
-import { AuthSignUp } from '../models/auth';
-import authSignUpSchema from '../validators/survivorRegistrationSchema';
+import { LEAFLET_DEFAULTS } from '../../constants/leafletDefaults';
+import { AuthSignUp } from '../../models/auth';
+import { useRegisterSurvivorMutation } from '../../services';
+import { buildSuccessToast } from '../../services/toastService';
+import authSignUpSchema from '../../validators/survivorRegistrationSchema';
+
 import SurvivorRegistrationContainer from './SurvivorRegistrationContainer';
-import { useRegisterSurvivorMutation } from '../services';
-import { buildSuccessToast } from '../services/toastService';
 
 const initialValues: AuthSignUp = {
   username: '',
