+++ src/constants/itemIconList.ts
@@ -5,7 +5,8 @@ import { ReactComponent as MedicineIcon } from '../images/icons/medicine.svg';
 import { ReactComponent as WaterIcon } from '../images/icons/water.svg';
 import { ReactComponent as WeaponIcon } from '../images/icons/weapon.svg';
 import { ReactComponent as UnknownIcon } from '../images/icons/unknown.svg';
-import { ItemIconMap, ItemIconType } from '../models/shared';
+
+import { ItemIconMap, ItemIconType } from '../models/item';
 
 export const ITEM_ICON_MAP: ItemIconMap = {
   [ItemIconType.Food]: FoodIcon,
